+++ archiso/profiledef.sh
@@ -49,4 +49,6 @@ file_permissions=(
   ["/usr/local/bin/holtos-btrfs-restore"]="0:0:755"
   ["/usr/local/bin/holtos-restore-snapshot"]="0:0:755"
   ["/usr/local/bin/holtos-snapshot-boot-notice"]="0:0:755"
+  ["/usr/local/bin/holtos-limine-other-os"]="0:0:755"
+  ["/usr/local/bin/holtos-rescan-bootmenu"]="0:0:755"
 )